backend/requirements.txt
alembic==1.18.4
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.1.4
click==8.3.1
cloudinary==1.44.1
Flask==3.1.2
flask-cors==6.0.2
# Flask-JWT==0.3.2 --- IGNORE ---
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.1
gunicorn==25.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.1
six==1.17.0
SQLAlchemy==2.0.46
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
